refactor: consolidate non-property-based merkle path tests into single file
- Combine merkle-path-parser.test.ts, merkle-path-parser.advanced.test.ts,
  and merkle-path-parser.arweave-compat.test.ts into one comprehensive test file
- Keep merkle-path-parser.property.test.ts separate as it uses fast-check
- Remove redundant test files after consolidation
- Maintain all test coverage and functionality
